Streamline pip install args
We have refactored the way `apache-superset` is installed.

First, we (optionally) specify the `$pip_index_url` separately. This
would overwrite the default (`false`, resolving to PyPI). Note that this
is a **breaking change** compared to the previous behaviour that was
adding the URL as an “extra index”, rather than “instead of PyPI”.

This enable us to simplify the parsing of `$pip_install_args` that is
then directly passed to the pip module (while concatenating the Array
into a space-separated String on the fly). We (graciously?) handle
the default case.

Signed-off-by: Étienne Boisseau-Sierra <etienne.boisseau-sierra@unipart.io>