Add OS dependencies for MySQL db
To be able to connect to MySQL dbs, one need to install the `mysql`
extras of the `apache-superset` Python package.

This extra dependency contains the `mysqlclient` package, which has OS
dependencies listed here: https://pypi.org/project/mysqlclient/

We make sure that these OS dependencies are installed (only if we
require MySQL).

Signed-off-by: Étienne Boisseau-Sierra <etienne.boisseau-sierra@unipart.io>